Add a scheduled workflow to figure out what's in the contexts
I would _love_ to know what information is available, to reply to
https://github.com/gitgitgadget-workflows/gitgitgadget-workflows/pull/2#discussion_r2305000995
which wants to know whether the information about the forkee is
available in the contexts of scheduled workflows in forked repositories.
This kind of information is surprisingly difficult to find on
docs.github.com, so let's just try it out.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>